feat: load duty page depending on vue.js
